@@ -10,57 +10,57 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~16.2.0",
-    "@angular/cdk": "^16.2.14",
-    "@angular/common": "~16.2.0",
-    "@angular/compiler": "~16.2.0",
-    "@angular/core": "~16.2.0",
-    "@angular/forms": "~16.2.0",
-    "@angular/material": "^16.2.14",
-    "@angular/platform-browser": "~16.2.0",
-    "@angular/platform-browser-dynamic": "~16.2.0",
-    "@angular/router": "~16.2.0",
-    "@nx/angular": "16.10.0",
+    "@angular/animations": "17.3.4",
+    "@angular/cdk": "17.3.4",
+    "@angular/common": "17.3.4",
+    "@angular/compiler": "17.3.4",
+    "@angular/core": "17.3.4",
+    "@angular/forms": "17.3.4",
+    "@angular/material": "17.3.4",
+    "@angular/platform-browser": "17.3.4",
+    "@angular/platform-browser-dynamic": "17.3.4",
+    "@angular/router": "17.3.4",
+    "@nx/angular": "18.2.4",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.13.0"
+    "zone.js": "0.14.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~16.2.0",
-    "@angular-devkit/core": "~16.2.0",
-    "@angular-devkit/schematics": "~16.2.0",
-    "@angular-eslint/eslint-plugin": "~16.0.0",
-    "@angular-eslint/eslint-plugin-template": "~16.0.0",
-    "@angular-eslint/template-parser": "~16.0.0",
-    "@angular/cli": "~16.2.0",
-    "@angular/compiler-cli": "~16.2.0",
-    "@angular/language-service": "~16.2.0",
-    "@nx/cypress": "16.10.0",
-    "@nx/eslint-plugin": "16.10.0",
-    "@nx/jest": "16.10.0",
-    "@nx/js": "16.10.0",
-    "@nx/linter": "16.10.0",
-    "@nx/web": "16.10.0",
-    "@nx/workspace": "16.10.0",
-    "@schematics/angular": "~16.2.0",
+    "@angular-devkit/build-angular": "17.3.4",
+    "@angular-devkit/core": "17.3.4",
+    "@angular-devkit/schematics": "17.3.4",
+    "@angular-eslint/eslint-plugin": "17.3.0",
+    "@angular-eslint/eslint-plugin-template": "17.3.0",
+    "@angular-eslint/template-parser": "17.3.0",
+    "@angular/cli": "~17.3.0",
+    "@angular/compiler-cli": "17.3.4",
+    "@angular/language-service": "17.3.4",
+    "@nx/cypress": "18.2.4",
+    "@nx/eslint": "18.2.4",
+    "@nx/eslint-plugin": "18.2.4",
+    "@nx/jest": "18.2.4",
+    "@nx/js": "18.2.4",
+    "@nx/web": "18.2.4",
+    "@nx/workspace": "18.2.4",
+    "@schematics/angular": "17.3.4",
     "@types/jest": "^29.4.0",
-    "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "^5.60.1",
-    "@typescript-eslint/parser": "^5.60.1",
-    "cypress": "^13.0.0",
-    "eslint": "~8.46.0",
-    "eslint-config-prettier": "8.1.0",
+    "@types/node": "^18.16.9",
+    "@typescript-eslint/eslint-plugin": "7.3.0",
+    "@typescript-eslint/parser": "7.3.0",
+    "cypress": "^13.6.6",
+    "eslint": "8.57.0",
+    "eslint-config-prettier": "9.0.0",
     "eslint-plugin-cypress": "^2.13.4",
     "husky": "^8.0.3",
     "jest": "^29.4.1",
     "jest-environment-jsdom": "^29.4.1",
-    "jest-preset-angular": "~13.1.0",
+    "jest-preset-angular": "14.0.3",
     "jsonc-eslint-parser": "^2.1.0",
     "lint-staged": "^15.2.2",
-    "nx": "16.10.0",
+    "nx": "18.2.4",
     "prettier": "^2.6.2",
     "ts-jest": "^29.1.0",
     "ts-node": "10.9.1",
-    "typescript": "~5.1.3"
+    "typescript": "5.4.5"
   }
 }
